+++ tenant/src/components/documents/tenant/Residency.vue
@@ -111,14 +111,14 @@ import TroubleshootingModal from "@/components/helps/TroubleshootingModal.vue";
     DocumentHelp,
     AllDeclinedMessages,
     NakedCard,
-    TroubleshootingModal
+    TroubleshootingModal,
   },
   computed: {
     ...mapGetters({
       user: "userToEdit",
-      tenantResidencyDocument: "getTenantResidencyDocument"
-    })
-  }
+      tenantResidencyDocument: "getTenantResidencyDocument",
+    }),
+  },
 })
 export default class Residency extends Vue {
   documents = DocumentTypeConstants.RESIDENCY_DOCS;
@@ -249,7 +249,7 @@ export default class Residency extends Vue {
 
   addFiles(fileList: File[]) {
     AnalyticsService.uploadFile("residency");
-    const nf = Array.from(fileList).map(f => {
+    const nf = Array.from(fileList).map((f) => {
       return { name: f.name, file: f, size: f.size };
     });
     this.files = [...this.files, ...nf];
@@ -263,7 +263,7 @@ export default class Residency extends Vue {
     this.uploadProgress = {};
     const fieldName = "documents";
     const formData = new FormData();
-    const newFiles = this.files.filter(f => {
+    const newFiles = this.files.filter((f) => {
       return !f.id;
     });
     if (!newFiles.length) return;
@@ -275,13 +275,14 @@ export default class Residency extends Vue {
       Vue.toasted.global.max_file({
         message: this.$i18n.t("max-file", [
           this.residencyFiles().length,
-          this.residencyDocument.maxFileCount
-        ])
+          this.residencyDocument.maxFileCount,
+        ]),
       });
+      this.files = [];
       return;
     }
 
-    Array.from(Array(newFiles.length).keys()).map(x => {
+    Array.from(Array(newFiles.length).keys()).map((x) => {
       const f: File = newFiles[x].file || new File([], "");
       formData.append(`${fieldName}[${x}]`, f, newFiles[x].name);
     });
@@ -297,7 +298,7 @@ export default class Residency extends Vue {
         this.fileUploadStatus = UploadStatus.STATUS_INITIAL;
         Vue.toasted.global.save_success();
       })
-      .catch(err => {
+      .catch((err) => {
         this.fileUploadStatus = UploadStatus.STATUS_FAILED;
         if (err.response.data.message.includes("NumberOfPages")) {
           Vue.toasted.global.save_failed_num_pages();
@@ -311,12 +312,12 @@ export default class Residency extends Vue {
   }
 
   residencyFiles() {
-    const newFiles = this.files.map(f => {
+    const newFiles = this.files.map((f) => {
       return {
         documentSubCategory: this.residencyDocument.value,
         id: f.name,
         name: f.name,
-        size: f.size
+        size: f.size,
       };
     });
     const existingFiles =
@@ -328,11 +329,11 @@ export default class Residency extends Vue {
 
   async remove(file: DfFile, silent = false) {
     AnalyticsService.deleteFile("residency");
-    if (file.id) {
+    if (file.path && file.id) {
       await RegisterService.deleteFile(file.id, silent);
     } else {
-      const firstIndex = this.files.findIndex(f => {
-        return f.name === file.name && f.file === file.file && !f.id;
+      const firstIndex = this.files.findIndex((f) => {
+        return f.name === file.name && !f.path;
       });
       this.files.splice(firstIndex, 1);
     }
@@ -341,4 +342,3 @@ export default class Residency extends Vue {
 </script>
 
 <style scoped lang="scss"></style>
-
